+++ flash/text/seq2seq/summarization/model.py
@@ -14,9 +14,9 @@
 from typing import Any, Dict, List, Optional
 
 import torch
+from torchmetrics import ROUGEScore
 
 from flash.core.utilities.types import LOSS_FN_TYPE, LR_SCHEDULER_TYPE, METRICS_TYPE, OPTIMIZER_TYPE
-from flash.text.seq2seq.core.metrics import RougeMetric
 from flash.text.seq2seq.core.model import Seq2SeqTask
 
 
@@ -72,8 +72,8 @@ class SummarizationTask(Seq2SeqTask):
             num_beams=num_beams,
             enable_ort=enable_ort,
         )
-        self.rouge = RougeMetric(
-            rouge_newline_sep=rouge_newline_sep,
+        self.rouge = ROUGEScore(
+            newline_sep=rouge_newline_sep,
             use_stemmer=use_stemmer,
         )